+++ src/index.js
@@ -10,9 +10,6 @@ import createSagaMiddleware from 'redux-saga';
 import rootWatcher from './services/saga/rootWatcher';
 
 
-
-
-
 const sagaMiddleware = createSagaMiddleware();
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 const store = createStore(reducer, composeEnhancers(applyMiddleware(sagaMiddleware)));
