+++ src/pages/CardPage.js
@@ -134,7 +134,9 @@ function CardPage () {
       dispatch(ontoggleReviewForm());
       document.body.style.overflow = "visible";
     }
+  
   }
+
 
   const quantityRewiews = reviews.length;
