+++ src/services/useHttp.hook.js
@@ -8,7 +8,6 @@ export const useHttp = () => {
        
         try {
             const response = await fetch(url, {method, body, headers});
-            //console.log(response)
 
             if (!response.ok) {
                 throw new Error(`Could not fetch ${url}, status: ${response.status}`);
@@ -30,23 +29,18 @@ export const useHttp = () => {
 
 export const useHttpPost = () => {
    
-    const requestPost = useCallback(async (url, body, method = 'POST',  headers = {'Content-Type': 'application/json;charset=utf-8'}) => {
+    const request = useCallback(async (url, body, method = 'POST',  headers = {'Content-Type': 'application/json;charset=utf-8'}) => {
 
         const bodyJSON = JSON.stringify(body)
-        console.log(bodyJSON)
         
         try {
             const response = await fetch(url, {method, body: bodyJSON, headers});
-            console.log(response)
                         
             if (!response.ok) {
                 throw new Error(`Could not fetch ${url}, status: ${response.status}`);
             }
 
-            const data = await response.json();
-            console.log(data)
-
-            return data;
+            return response;
         } catch(e) {
             
             throw e;
@@ -55,5 +49,5 @@ export const useHttpPost = () => {
 
     
 
-    return {requestPost}
+    return {request}
 }
